react-app: Move YourStake model to validator models
refs #309
